Ember Interface
A new interface was add to the project
All the interface was implemented in ember
This is the virst version, the interface may send the url and get the short code, besides it shows the short_code or errors if it exists. Finally, the interface show the 100 most accessed urls.